# lib/api-zod/src/generated/types/poseMatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KINECTRA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PoseMatch {
  matchName: string;
  similarity: number;
  role: string;
  description: string;
  idealVector: number[];
}