# lib/api-zod/src/generated/types/sessionEndInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KINECTRA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Snapshot } from './snapshot';

export interface SessionEndInput {
  frameCount: number;
  avgPostureScore: number;
  avgAlignmentScore: number;
  avgStabilityScore: number;
  avgEfficiencyScore: number;
  overallScore: number;
  warnings: string[];
  snapshots?: Snapshot[];
}