# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KINECTRA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SessionInputAnalysisType } from './sessionInputAnalysisType';
import type { SessionInputDominantHand } from './sessionInputDominantHand';
import type { SessionInputSkillLevel } from './sessionInputSkillLevel';

export interface SessionInput {
  athleteName: string;
  analysisType: SessionInputAnalysisType;
  skillLevel: SessionInputSkillLevel;
  dominantHand: SessionInputDominantHand;
}